agents/crewai-weather/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pyyaml==6.0.3
requests==2.32.5
crewai==0.201.1
crewai-tools==0.75.0